nous avons fini de mettre en place toute la partie php avec aussi la vérification de tout les parametres et le celui qui a été le plus dur la pagination pour toutes les pages en backend et en frontend nous allons maintent passer à la séparation du javascript du css tout en incorporant le bootstrap

+++ controller/postBackend.php
@@ -4,10 +4,25 @@ require('model/Post.php');
 function postCreate($title, $content){
 
 	$post = new Post($title, $content);
+	
+try{ 	
+	if(!empty($title) && !empty($content) && strlen($title)<255 )
+		{
+			$post->create();
+			header('Location:index.php?action=postListView');
+		}
+	else
+		{
+			throw new Exception("Vous devez remplir tous les champs.");
+			
+		}
+	}
+catch(Exception $e)
+	{
+		$errorMessage = 'Erreur : '.$e->getMessage();
+		require('view/errorView.php');
+	}
 
-	$post->create();
-
-	header('Location:index.php?action=postListView');
 
 }
 
@@ -16,28 +31,44 @@ function postCreationView(){
 	require('view/postCreationView.php');
 }
 
-function postListView (){
+function postListView ($page, $postPerPage){
 
 	$post = new Post("init", "init");
+try{ 
+	if( $page>0  && $postPerPage>0)
+		{
+			$postList = $post->postList($page, $postPerPage);
+			require('view/backendPostListView.php');
+		}
+	else
+		{
+			throw new Exception("La requete n'a pas pu être pris en compte.");
+			
+		}
+	}
+catch(Exception $e)
+	{
+		$errorMessage = 'Erreur : '.$e->getMessage();
+		require('view/errorView.php');
+	}
+	
 
-	$postList = $post->postList();
-
-	require('view/backendPostListView.php');
+	
 
 }
 
-function postDelete($postId){
+function postDelete($postId, $page){
 
 	$post = new Post("init", "init");
 
 	$check = $post->checkPost($postId);
 
 try{
-		if($check)
+		if($check && $page>0)
 			{
 				$post->delete($postId);
 
-				header('Location:index.php?action=postListView');
+				header('Location:index.php?action=postListView&page='.$page);
 			}
 		else
 			{
@@ -83,18 +114,18 @@ catch(Exception $e)
 
 }
 
-function postUpdate($title, $content, $postId){
+function postUpdate($title, $content, $postId, $page){
 
 	$post = new Post($title, $content);
 
 	$check = $post->checkPost($postId);
 
 try{
-	if($check)
+	if($check && !empty($title) && !empty($content) && !empty($postId) && strlen($title)<255  && $page>0)
 		{
 			$post->update($postId);
 
-			header('Location:index.php?action=postListView');
+			header('Location:index.php?action=postListView&page='.$page);
 		}
 	else
 		{
